PY3: Ensure sys.ext is only ever passed an integer

+++ tests/System/GridTestSubmission/exe-script-with-input-and-ancestor.py
@@ -17,4 +17,4 @@ import sys
 # Main
 if __name__ == '__main__':
 
-  sys.exit(system('''ls 081616_0000000213.raw''') / 256)
+  sys.exit(int(system('''ls 081616_0000000213.raw''') / 256))
