+++ tests/System/GridTestSubmission/exe-script-with-input-jenkins.py
@@ -17,4 +17,4 @@ from os import system
 # Main
 if __name__ == '__main__':
 
-  sys.exit(system('''cat jenkinsInputTestFile.txt''') / 256)
+  sys.exit(int(system('''cat jenkinsInputTestFile.txt''') / 256))
