+++ tests/System/GridTestSubmission/exe-script-with-input-single-location.py
@@ -17,4 +17,4 @@ import sys
 # Main
 if __name__ == '__main__':
 
-  sys.exit(system('''cat testInputFileSingleLocation.txt''') / 256)
+  sys.exit(int(system('''cat testInputFileSingleLocation.txt''') / 256))
